Add Article, Author, and Subject classes with properties
Created  `Article.cs` to include an `Article` class inheriting from `AuditableEntity` with properties for `Title`, `Content`, `Image`, `ViewCount`, and `ReadTime`, and navigation properties for `AuthorId`, `Author`, `SubjectId`, and `Subject`.

Created  `Author.cs` to include an `Author` class inheriting from `BaseUser` with a navigation property for a collection of `Articles`.

Created  `Subject.cs` to include a `Subject` class inheriting from `AuditableEntity` with a property for `Name` and a navigation property for a collection of `Articles`.